+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 Changelog is kept with respect to version 0.11 of Entropies.jl. From version v2.0 onwards, this package has been renamed to ComplexityMeasures.jl.
 
+## 3.3
+
+- Added the `SequentialPairDistances` outcome space. In the literature, this outcome
+    space has been used to compute the "distribution entropy", which can be reproduced
+    with `information(Shannon(), SequentialPairDistances(x), x)`. It can of course also 
+    be used in combination with any other information measure.
+- Added the `PairDistanceEncoding` encoding.
+- Added the `entropy_distribution` convenience function.
+
 ## 3.2
 
 - `missing_outcomes` only works with count-based outcome spaces, which is what it should be doing based on its conceptual definition. Previous signature has been deprecated.